Moved operation enums into separate module
