+++ lava/lib/dnf/operations/enums.py
@@ -1,0 +1,18 @@
+# Copyright (C) 2021 Intel Corporation
+# SPDX-License-Identifier: BSD-3-Clause
+# See: https://spdx.org/licenses/
+
+from enum import Enum, unique, auto
+
+
+@unique
+class ReduceMethod(Enum):
+    """Enum for reduce methods of ReduceDims Operation"""
+    SUM = auto()  # ReduceDims will sum all synaptic weights of collapsed dim
+    MEAN = auto()  # ReduceDims will compute mean of weights of collapsed dim
+
+    @classmethod
+    def validate(cls, reduce_method):
+        """Validate <reduce_op> against all valid enum values"""
+        if not isinstance(reduce_method, ReduceMethod):
+            raise TypeError("reduce_method must be of value ReduceMethod")
